+++ pkg/pathsync/plan.go
@@ -10,7 +10,7 @@ type Plan struct {
 	Engine                Engine
 	PreserveSourceDirName bool
 	Mirror                bool
-	DisableSafeCopy       bool
+	SafeCopy              bool
 
 	RetryCount        int
 	RetryWait         time.Duration